Extract the table filters into a common hook
